// layers/editor/types/context.model.ts
export type ReactZoomPanPinchContext = typeof ZoomPanPinch.prototype

export interface ReactZoomPanPinchContextState {
  instance: ReactZoomPanPinchContext
  state: ReactZoomPanPinchState
}

export type ReactZoomPanPinchContentRef = {
  instance: ReactZoomPanPinchContext
} & ReactZoomPanPinchHandlers

export type ReactZoomPanPinchRef = ReactZoomPanPinchContextState &
  ReactZoomPanPinchHandlers

export interface ReactZoomPanPinchState {
  previousScale: number
  scale: number
  positionX: number
  positionY: number
}

export interface ReactZoomPanPinchHandlers {
  zoomIn: ReturnType<typeof zoomIn>
  zoomOut: ReturnType<typeof zoomOut>
  setTransform: ReturnType<typeof setTransform>
  resetTransform: ReturnType<typeof resetTransform>
  centerView: ReturnType<typeof centerView>
  zoomToElement: ReturnType<typeof zoomToElement>
}